EpilepsyDetection/streamlit-dashboard/requirements.txt
joblib==1.4.2
keras==3.3.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
pyradiomics==3.1.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
scipy==1.13.0
SimpleITK==2.3.1
SimpleITK==2.3.1
streamlit==1.34.0
streamlit_authenticator==0.3.2
streamlit_drawable_canvas==0.9.2
tensorflow==2.16.1
